Fix editing links in the minibuffer
Apparently, something was changed in Org-mode recently.

Consider a [[type:body]] citation link.  If the point is inside "body",
then 'org-ref-cite-keymap' is active but if the point is inside "type",
then 'org-mouse-map' is active.  Since prettification uses 'display'
property, which uses the first character of the hidden text, it follows
that 'org-mouse-map' is active for prettified links.  So we have to bind
keys there.

* org-ref-prettify.el (org-ref-prettify-edit-link-at-point-maybe)
(org-ref-prettify-edit-link-at-mouse-maybe): New commands.
Bind them in 'org-mouse-map' instead of 'org-ref-cite-keymap'.